requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-filter==21.1
djangorestframework-simplejwt==5.2.0
drf-ratelimit==3.0.0
